v0.4a
Dress-up scene *should* be fully functional. The latency bonus *should* reflect in the actual rhythm game too.

+++ Assets/Scripts/DressUpStatBonuses.cs
@@ -5,6 +5,6 @@ using UnityEngine;
 public static class DressUpStatBonuses
 {
     public static double scoreMultiplier {get; set;}
-    public static int leniency {get; set;}
+    public static double leniency {get; set;}
     public static int scoreThreshold {get; set;}
 }
